+++ src/GameLogic.hpp
@@ -44,7 +44,7 @@ int SceneInBox(D3DVECTOR point);
 int CalculateView(XMFLOAT3 EyeBall, XMFLOAT3 LookPoint, float angle, bool distancecheck);
 int CalculateViewMonster(XMFLOAT3 EyeBall, XMFLOAT3 LookPoint, float angle, float angy);
 void PlayerNonIndexedBox(int pmodel_id, int curr_frame, int angle, float wx, float wy, float wz, int monsterid);
-void PlayerIndexedBox(int pmodel_id, int curr_frame, int angle, float wx, float wy, float wz);
+void PlayerIndexedBox(int pmodel_id, int curr_frame, float angle, float wx, float wy, float wz);
 void MakeBoundingBox();
 int FindGunTexture(char *p);
 void PlayerToD3DVertList(int pmodel_id, int curr_frame, float angle, int texture_alias, int tex_flag, float xt, float yt, float zt, int nextFrame = -1);
